Splitting RBR code into own files.
